Added sound as reaction to button click. Added function that pixelfies distance diff.
Currently the button shows up when facing certain direction (that was only because it was easier to test than GPS changes).
It disappears then out of that span.
It does not show up again after the sound clip has been played once - so the option of replaying within a page reload has been eliminated. We need to fix that somehow, in case someone misses a part of the story.

The function that pixelfies just takes GPS coordinates of "place" and compares with "you", then converts that distance in accordance with degrees per km in Siglufjörður, and finally scales that up.